back-logs/requirements/base.txt
#fasapi
python-decouple==3.8
aiofiles==23.2.1
pydantic==2.6.4
fastapi==0.110.2
python-dotenv==1.0.1
zeep==4.2.1
python-multipart==0.0.9
pandas==2.2.1
matplotlib==3.8.4
httpx==0.27.0 #HTTP client

#mongo
dnspython==2.6.1
motor==3.6.0
typing-extensions==4.10.0
pathspec==0.12.1

#server 
uvicorn==0.29.0